requirements.txt
pandas 
mlflow==2.11.2
dvc
notebook
numpy
scikit-learn
catboost
xgboost
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
kaggle
requests
-e .